Changes according to the comments
